+++ tests/Feature/UpdatePasswordTest.php
@@ -2,11 +2,11 @@
 
 namespace Tests\Feature;
 
-use App\Models\User;
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use Illuminate\Support\Facades\Hash;
 use Laravel\Jetstream\Http\Livewire\UpdatePasswordForm;
 use Livewire\Livewire;
+use Modules\User\Entities\User;
 use Tests\TestCase;
 
 class UpdatePasswordTest extends TestCase